backend/pyproject.toml
[tool.poetry]
name = "restaurant-reservation-api"
version = "2.0.0"
description = "FastAPI-based restaurant reservation system"
authors = ["Your Name <you@example.com>"]

[[tool.poetry.packages]]
include = "app"
from = "."

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
firebase-admin = "^7.1.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
requests = "^2.32.5"
python-dotenv = "^1.1.1"
python-dateutil = "^2.9.0.post0"
email-validator = "^2.1.1"

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.28.1"
black = "^24.3.0"
ruff = "^0.3.4"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
